Make serde a feature, but include it as a default feature
