+++ README.md
@@ -21,7 +21,27 @@ For taking advantage of simdjson your system needs to be SIMD compatible. This m
 
 If you are writing perormance centric code, make sure to use jemalloc and not the system allocator (that has now become default in rust), it gives an very noticable boost imperformance.
 
+## serde
 
+simd-json.rs is compatible with serde and serde-json. The Value types provided implement serializers and deserialisers. In addition to that simd-json.rs implements the Deserializer for the parser so it can deserialize anything that implements the serde Deserialize trait.
+
+That said serde is contained in the `serde-comp` feature which is part of the default feature set, but it can be disabled.
+
+### serializing
+
+sims-json.rs does not provide capabiltiy to serialize JSON data, serde-json does a good job at this and there would be very little won by re-implementing it.
+
+### unsafe
+
+simd-json.rs uses **a lot** of unsafe code first of all since all SIMD-intrinsics are inherently unsafe and also to work around some bottlenecks safe rust code. So this requires extra scrutiny and do to this diligently testing takes 5 shapes:
+
+* Poor old unit tests - to test general 'obvious' cases and edge cases as well as regressions
+* Property based testing on valid json - tests if valid but random generated json parses the same in simd-json and in serde-json (floats here are excloded since slighty different parsing algorihtms lead to slighty different results)
+* Property based testing on random 'human readable' data - make sure that randomly generated sequences of pritnable characters don't panic or crash the parser (they might and often error so - they are not valid json!)
+* Property based testing on random byte sequences - make sure that no random set of bytes will crash the parser
+* Fuzzing (using afl) - fuzz based on upstream simd pass/fail cases
+
+This sure doens't ensure complete safetly but it does go a long way.
 
 ## Other interesting things
